:gear: un début de propreté dans les mécanismes
POC de mécanisme dznsnun style nouveau et dans un nouveau fichier.

reste à abstraire la fonction v
